feat: concatenating Azure Openai API

+++ AI_customer_service.py
@@ -18,31 +18,43 @@ from distance_calc import StoreDistanceEvaluator
 class ChatGPT_AVD:
     def __init__(self):
         self.OPEN_AI_KEY_DICT = eval(os.getenv('OPENAI_API_KEY'))
+        self.AZURE_OPENAI_CONFIG = eval(os.getenv('AZURE_OPENAI_CONFIG'))
 
     def get_keys(func):
         def inner(self, message, model="gpt-3.5-turbo", timeout=60):
+            config = self.AZURE_OPENAI_CONFIG
             # get token_id
-            query = 'SELECT id, counts FROM web_push.AI_service_token_counter x ORDER BY counts limit 1;'
             if 'gpt-4' in model:
+                query = 'SELECT id, counts FROM web_push.AI_service_token_counter x ORDER BY counts limit 1;'
                 query = 'x WHERE id < 6'.join(query.split('x'))
-            token_id = DBhelper('jupiter_new').ExecuteSelect(query)[0][0]
+                token_id = DBhelper('jupiter_new').ExecuteSelect(query)[0][0]
+                config = {'api_key': self.OPEN_AI_KEY_DICT[token_id],
+                          'api_type': 'open_ai',
+                          'api_base': 'https://api.openai.com/v1',
+                          'api_version': None,
+                          'kwargs': {'model': model}}
+                # update token counter
+                DBhelper('jupiter_new').ExecuteDelete(f'UPDATE web_push.AI_service_token_counter SET counts = counts + 1 WHERE id = {token_id}')
 
-            # update token counter
-            DBhelper('jupiter_new').ExecuteDelete(f'UPDATE web_push.AI_service_token_counter SET counts = counts + 1 WHERE id = {token_id}')
             try:
-                res = func_timeout(timeout, func, (self, message, token_id, model))
+                res = func_timeout(timeout, func, (self, message, config))
             except:
                 res = 'timeout'
-            DBhelper('jupiter_new').ExecuteDelete(f'UPDATE web_push.AI_service_token_counter SET counts = counts - 1 WHERE id = {token_id}')
+
+            if 'gpt-4' in model:
+                DBhelper('jupiter_new').ExecuteDelete(f'UPDATE web_push.AI_service_token_counter SET counts = counts - 1 WHERE id = {token_id}')
             return res
         return inner
 
     @get_keys
-    def ask_gpt(self, message, token_id=None, model="gpt-3.5-turbo"):
-        openai.api_key = self.OPEN_AI_KEY_DICT[token_id]
-        if type(message) == str:
-            message = [{'role': 'user', 'content': message}]
-        completion = openai.ChatCompletion.create(model=model, messages=message)
+    def ask_gpt(self, message, config):
+        openai.api_key = config.get('api_key')
+        openai.api_type = config.get('api_type')
+        openai.api_base = config.get('api_base')
+        openai.api_version = config.get('api_version')
+        kwargs = config.get('kwargs')
+        kwargs['messages'] = [{'role': 'user', 'content': message}] if type(message) == str else message
+        completion = openai.ChatCompletion.create(**kwargs)
         return completion['choices'][0]['message']['content']
 
     def num_tokens_from_messages(self, messages, model="gpt-3.5-turbo"):
@@ -118,7 +130,7 @@ class ChatGPT_AVD:
                 if v.get('pagemap') and v.get('pagemap').get('metatags') and v.get('pagemap').get('metatags')[0].get('og:description'):
                     chatgpt_query += f""",description = {v.get('pagemap').get('metatags')[0].get('og:description')}" """
             # chatgpt_query += f"""\n\n\nCurrent date: {date}\n\nInstructions: If you are "{web_id_conf['web_name']}" customer service. Using the information of results or following the flow of conversation, write a comprehensive reply to the given query in 繁體中文 and following the rules below:\nAlways cite the information from the provided results using the [number] notation in the end of that sentence.\nWrite Bullet list for each subject if you recommend products.\n"親愛的顧客您好，" in the beginning.\n"祝您愉快！" in the end.\n\nQuery: {message}"""
-            chatgpt_query += f"""\n\n\nCurrent date: {date}\n\nInstructions: As a customer service representative for "{web_id_conf['web_name']}". Your task is to respond to the Query below in 繁體中文. Always start with "親愛的顧客您好，" and end with "祝您愉快！". Your should first address the customer's question. Then, recommend three products by using bullet points. Ensure that your response is comprehensive, helpful and response is generated entirely from the information provided. Use the [number] notation to cite your sources after each sentence.\n\nCustomer's Query: {message}"""
+            chatgpt_query += f"""\n\n\nCurrent date: {date}\n\nInstructions: As a customer service representative for "{web_id_conf['web_name']}". Your task is to respond to the Query below in 繁體中文. Always start with "親愛的顧客您好，" and end with "祝您愉快！". Your should first address the customer's question. Then, recommend three products by using bullet points. Ensure that your response is comprehensive, helpful and response is generated entirely from the information provided. Use the [number] notation to cite your sources at the end of each sentence.\n\nCustomer's Query: {message}"""
         else:
             chatgpt_query = f"""\n\n\nCurrent date: {date}\n\nInstructions: If you are the brand, "{web_id_conf['web_name']}"({web_id_conf['web_id']}) customer service and there is no search result in product list, write a comprehensive reply to the given query. Reply in 繁體中文 and Following the rule below:\n"親愛的顧客您好，" in the beginning.\n"祝您愉快！" in the end.\n\nQuery: {message}"""
         gpt_query += [{'role': 'user', 'content': chatgpt_query}]
@@ -415,7 +427,7 @@ class QA_api:
             return "親愛的顧客您好，您的提問長度超過限制，請縮短問題後重新發問。"
 
         #types, orders = self.get_order_type(web_id, user_id, message)
-
+      
 
 
         history_df = self.get_history_df(web_id, info)
@@ -423,22 +435,6 @@ class QA_api:
         self.logger.print('QA歷史紀錄:\n', history)
         flags, f = self.judge_question_type(message)
         self.logger.print(f'客戶意圖:\t{flags}\n{f}')
-        if gps_location:
-            store_result = self.search_nearest_store_nineyi000360(gps_location, history)
-            if store_result:
-                flags['store_address'] = True
-                flags['hot'] = True
-                message = '請提供result的全家便利商店的位置資訊。'
-            else:
-                return "親愛的顧客您好，我們不確定您的問題或需求，如果您有任何疑慮或需要任何協助，請隨時聯絡我們的客戶服務團隊。"
-        # if types == 1:
-        #     gpt_query = self.ChatGPT.get_gpt_order_query(orders,message)
-        #     self.logger.print('訂單系統ChatGPT輸入:\n',gpt_query)
-        #     gpt_answer = translation_stw(self.ChatGPT.ask_gpt(gpt_query, timeout=60)).replace('，\n', '，')
-        #     if '生活愉快！' in gpt_answer:
-        #         gpt_answer = '生活愉快！'.join(gpt_answer.split("生活愉快！")[:-1]) + '生活愉快！'
-        #     self.logger.print('訂單系統ChatGPT輸出:\n', gpt_query)
-        #     return gpt_answer
 
         if gps_location:
             store_result = self.search_nearest_store_nineyi000360(gps_location, history)
@@ -448,7 +444,18 @@ class QA_api:
                 message = '給我全家便利商店的位置資訊。'
             else:
                 return "親愛的顧客您好，我們不確定您的問題或需求，如果您有任何疑慮或需要任何協助，請隨時聯絡我們的客戶服務團隊。"
+                
+        # if types == 1:
+        #     gpt_query = self.ChatGPT.get_gpt_order_query(orders,message)
+        #     self.logger.print('訂單系統ChatGPT輸入:\n',gpt_query)
+        #     gpt_answer = translation_stw(self.ChatGPT.ask_gpt(gpt_query, timeout=60)).replace('，\n', '，')
+        #     if '生活愉快！' in gpt_answer:
+        #         gpt_answer = '生活愉快！'.join(gpt_answer.split("生活愉快！")[:-1]) + '生活愉快！'
+        #     self.logger.print('訂單系統ChatGPT輸出:\n', gpt_query)
+        #     return gpt_answer
+        
         if not gps_location and flags.get('store_address') and web_id == 'nineyi000360':
+     
             keyword = ''
             gpt_query = [{'role': 'user', 'content': message}]
             gpt_answer = answer = "親愛的顧客您好，若是需要查詢最近的全家便利店位置，請提供我們您現在的位置。"
